Move ssd large test into tests/accuracy/
